requirements.txt
adal==1.2.7
azure-common==1.1.27
azure-core==1.18.0
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.1.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.1.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azureml-core==1.34.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
contextlib2==21.6.0
cryptography==3.4.8
databases==0.5.1
DateTime==4.3
docker==5.0.2
fields==5.0.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.1
jmespath==0.10.0
jsonify==0.5
jsonpickle==2.0.0
lazyload==0.0.2
MarkupSafe==2.0.1
marshmallow==3.13.0
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
oauthlib==3.1.1
orm==0.1.8
pathspec==0.9.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2021.1
pywin32==227
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.6
schema==0.7.4
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==1.4.23
typesystem==0.2.4
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
WTForms==2.3.3
zope.interface==5.4.0
